Get rid of obsolescent AC_HEADER_TIME. Cleanup about sys/time.h usage.
The AC_HEADER_TIME macro is obsolescent in autoconf 2.69 (2012),
and autoconf 2.71 (2021) outputs a warning, which makes autogen.sh
fail due to "--warnings=all,error", so let's avoid the warnings.
This macro defines TIME_WITH_SYS_TIME, used only in tests/tests.c
(but actually not needed). We actually have two cases:
  1. HAVE_GETTIMEOFDAY is defined: we use gettimeofday(), which
     needs <sys/time.h>.
  2. Otherwise: we use time(), which needs <time.h>.

Based on that, we can simplify the condition in tests/tests.c for
the inclusion of time-related headers, and we no longer need to
check the <sys/time.h> availability. If gettimeofday() is defined
but <sys/time.h> is not available (which shouldn't occur), then
the current gettimeofday() usage may be incorrect anyway.

Changes:
  * acinclude.m4:
      - no longer use AC_HEADER_TIME;
      - no longer check the sys/time.h header.
  * tests/tests.c: simplify the inclusion of time-related headers.