[src/mpfr-impl.h] Added cast to int for __LINE__ in formatted output.
The standard just says that __LINE__ in an integer constant. So its
type might be larger than int, even though the value is representable
in an int (like in MPFR). This concerns logging code, only available
for GCC compatible compilers, so that there should not be any issue
without the cast, but just in case...